move schemas to lib
