+++ src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from 'src/schema';
+import { User, UserSchema } from 'src/lib/schema';
 
 @Module({
   imports: [
